chore(core): convert ember/index.d.ts to .ts
